LPS-55433 Move integration tests related to the startup action to the test module
