+++ modules/apps/document-library/document-library-test/test/integration/com/liferay/document/library/events/test/dependencies/ddmstructure.xml
